add item to start works
